make root a functional component
