+++ web/package.json
@@ -44,6 +44,7 @@
     "@types/node": "^18.7.13",
     "@types/react": "^18.0.17",
     "@types/react-dom": "^18.0.6",
+    "@types/react-helmet": "5",
     "@types/react-modal": "^3.11.2",
     "@types/react-router": "^5.1.18",
     "@types/react-router-dom": "^5",
