+++ web/src/index.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { createRoot } from "react-dom/client";
 import ReplicatedErrorBoundary from "./components/shared/ErrorBoundary";
-import Root from "./Root";
+import { Root } from "./Root";
 
 const container = document.getElementById("app");
 const root = createRoot(container); // createRoot(container!) if you use TypeScript
