+++ web/src/routes/RootRoutes.tsx
@@ -1,0 +1,17 @@
+// import React from "react";
+
+const RootRoutes: [] = [
+  // {
+  //   path: '/app',
+  //   element: <App />,
+  //   children: [
+  //     { path: '/discussions/*', element: <DiscussionsRoutes /> },
+  //     { path: '/users', element: <Users /> },
+  //     { path: '/profile', element: <Profile /> },
+  //     { path: '/', element: <Dashboard /> },
+  //     { path: '*', element: <Navigate to="." /> },
+  //   ],
+  // },
+];
+
+export { RootRoutes };
